superior-trading-bot-phases/phase2-intelligence/requirements.txt
numpy>=1.24.0
scipy>=1.11.0
scikit-learn>=1.3.0
umap-learn>=0.5.4
hdbscan>=0.8.33
xgboost>=2.0.0
pandas>=2.1.0
